apps/backend/requirements.txt
fastapi
uvicorn[standard]
mangum
requests
httpx
yfinance
praw
python-decouple
python-dotenv
